refactor: organize PlansPanel component
