+++ packages/manager/src/features/Databases/DatabaseCreate/DatabaseCreate.tsx
@@ -46,7 +46,7 @@ import { databaseEngineMap } from 'src/features/Databases/DatabaseLanding/Databa
 import { enforceIPMasks } from 'src/features/Firewalls/FirewallDetail/Rules/FirewallRuleDrawer.utils';
 import PlansPanel, {
   PlanSelectionType,
-} from 'src/features/Linodes/LinodesCreate/SelectPlanPanel/PlansPanel';
+} from 'src/features/components/PlansPanel/PlansPanel';
 import { typeLabelDetails } from 'src/features/Linodes/presentation';
 import { useFlags } from 'src/hooks/useFlags';
 import {
