+++ packages/manager/src/features/Support/SupportTickets/SupportTicketDialog.tsx
@@ -13,6 +13,7 @@ import { Accordion } from 'src/components/Accordion';
 import { ActionsPanel } from 'src/components/ActionsPanel/ActionsPanel';
 import { Dialog } from 'src/components/Dialog/Dialog';
 import Select, { Item } from 'src/components/EnhancedSelect/Select';
+import { Link } from 'src/components/Link';
 import { Notice } from 'src/components/Notice/Notice';
 import { EntityForTicketDetails } from 'src/components/SupportLink/SupportLink';
 import { TextField } from 'src/components/TextField';
@@ -31,6 +32,7 @@ import {
   getErrorMap,
   getErrorStringOrDefault,
 } from 'src/utilities/errorUtils';
+import { reduceAsync } from 'src/utilities/reduceAsync';
 import scrollErrorIntoView from 'src/utilities/scrollErrorIntoView';
 import { storage } from 'src/utilities/storage';
 
@@ -44,8 +46,6 @@ import SupportTicketSMTPFields, {
   smtpDialogTitle,
   smtpHelperText,
 } from './SupportTicketSMTPFields';
-import { Link } from 'src/components/Link';
-import { reduceAsync } from 'src/utilities/reduceAsync';
 
 const useStyles = makeStyles((theme: Theme) => ({
   expPanelSummary: {
