+++ packages/manager/src/features/components/PlansPanel/PlanInformation.test.tsx
